ccproxy/requirements.txt
# Core web framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for Claude API
httpx==0.25.2

# Authentication and security
PyJWT==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Performance and caching
redis==5.0.1
aioredis==2.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-test==0.22.0

# Type checking
mypy==1.7.1
types-redis==4.6.0.11

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Documentation
pydantic[email]==2.5.0